New concept of countable events introduced for task scheduler. So far, a task received a task parameter, which was set by the instance activating the task. Now, as an alternative, the activating task can coherently communicate to the activated task, how often it has sent a signal. Producer-consumer patterns are supported; the consumer will be awoken by every produced entity and it'll know exactly, how many entities it receives from the producer. The use of bit masks allows conveying different events to the task and the task is activated by any of them

+++ samples/safe-RTOS/code/application/basicTest/Z4A/prf_processFailureInjection.h
@@ -315,12 +315,12 @@ extern prf_cmdFailure_t SHARED(prf_cmdFailure);
  */
 
 /** User task intenionally producing severe errors, which would crash an unprotected RTOS. */
-int32_t prf_taskInjectError(uint32_t PID, uintptr_t taskParam);
+int32_t prf_taskInjectError(uint32_t PID, uint32_t taskParam);
 
 /** User task in same process, running at lower priority. */
-int32_t prf_task17ms(uint32_t PID, uintptr_t taskParam);
+int32_t prf_task17ms(uint32_t PID, uint32_t taskParam);
 
 /** User task in same process, running at higher priority. */
-int32_t prf_task1ms(uint32_t PID, uintptr_t taskParam);
+int32_t prf_task1ms(uint32_t PID, uint32_t taskParam);
 
 #endif  /* PRF_PROCESSFAILUREINJECTION_INCLUDED */
