+++ samples/safe-RTOS/code/application/triCore/Z4A/prr_processReporting.c
@@ -163,7 +163,7 @@ int32_t prr_taskReportFailure( uint32_t PID ATTRIB_UNUSED
  *   @param taskParam
  * A variable task parameter. Here not used.
  */
-int32_t prr_taskReporting(uint32_t PID ATTRIB_UNUSED, uintptr_t taskParam ATTRIB_UNUSED)
+int32_t prr_taskReporting(uint32_t PID ATTRIB_UNUSED, uint32_t taskParam ATTRIB_UNUSED)
 {
     const unsigned int cpuLoadZ4A = syc_cpuLoadZ4A
                      , cpuLoadZ4B = m4b_cpuLoadZ4B
@@ -283,7 +283,7 @@ int32_t prr_taskReporting(uint32_t PID ATTRIB_UNUSED, uintptr_t taskParam ATTRIB
  * A variable task parameter. Here not used.
  */
 int32_t prr_taskTestContextSwitches( uint32_t PID ATTRIB_UNUSED
-                                   , uintptr_t taskParam ATTRIB_UNUSED
+                                   , uint32_t taskParam ATTRIB_UNUSED
                                    )
 {
     /* The next call produces 100 * noCycles*(waitTimePerCycleInUS/1000) / 11 percent of
